# Add accesslog async writer