--- a/middlewares/accesslog/logger.go
+++ b/middlewares/accesslog/logger.go
@@ -3,6 +3,7 @@ package accesslog
 import (
 	"context"
 	"fmt"
+	"io"
 	"net"
 	"net/http"
 	"net/url"
@@ -33,12 +34,18 @@ const (
 
 // LogHandler will write each request and its response to the access log.
 type LogHandler struct {
-	config              *types.AccessLog
-	logger              *logrus.Logger
-	asyncWriter         *asyncWriter
-	asyncWriterChanSize int64
-	mu                  sync.Mutex
-	httpCodeRanges      types.HTTPCodeRanges
+	config            *types.AccessLog
+	logger            *logrus.Logger
+	logWriter         logWriter
+	logWriterChanSize int64
+	filePath          string
+	mu                sync.Mutex
+	httpCodeRanges    types.HTTPCodeRanges
+}
+
+type logWriter interface {
+	Close() error
+	io.Writer
 }
 
 // NewLogHandler creates a new LogHandler
@@ -63,20 +70,25 @@ func NewLogHandler(config *types.AccessLog) (*LogHandler, error) {
 		return nil, fmt.Errorf("unsupported access log format: %s", config.Format)
 	}
 
-	asyncWriter := newAsyncWriter(config.AsyncWriterChanSize, file)
 	logger := &logrus.Logger{
-		Out:       asyncWriter,
 		Formatter: formatter,
 		Hooks:     make(logrus.LevelHooks),
 		Level:     logrus.InfoLevel,
 	}
 
 	logHandler := &LogHandler{
-		config:              config,
-		logger:              logger,
-		asyncWriter:         asyncWriter,
-		asyncWriterChanSize: config.AsyncWriterChanSize,
+		config: config,
+		logger: logger,
 	}
+
+	var lw logWriter = file
+	if config.Async != nil {
+		lw = newAsyncWriter(config.Async.AsyncWriterChanSize, file)
+		logHandler.logWriterChanSize = config.Async.AsyncWriterChanSize
+	}
+
+	logger.Out = lw
+	logHandler.logWriter = lw
 
 	if config.Filters != nil {
 		if httpCodeRanges, err := types.NewHTTPCodeRanges(config.Filters.StatusCodes); err != nil {
@@ -164,7 +176,7 @@ func (l *LogHandler) ServeHTTP(rw http.ResponseWriter, req *http.Request, next h
 
 // Close closes the Logger (i.e. the file etc).
 func (l *LogHandler) Close() error {
-	return l.asyncWriter.Close()
+	return l.logWriter.Close()
 }
 
 // Rotate closes and reopens the log file to allow for rotation
@@ -172,20 +184,20 @@ func (l *LogHandler) Close() error {
 func (l *LogHandler) Rotate() error {
 	var err error
 
-	if l.asyncWriter != nil {
-		defer func(f *asyncWriter) {
+	if l.logWriter != nil {
+		defer func(f logWriter) {
 			f.Close()
-		}(l.asyncWriter)
+		}(l.logWriter)
 	}
 
 	file, err := os.OpenFile(l.config.FilePath, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0664)
 	if err != nil {
 		return err
 	}
-	l.asyncWriter = newAsyncWriter(l.asyncWriterChanSize, file)
+	l.logWriter = newAsyncWriter(l.logWriterChanSize, file)
 	l.mu.Lock()
 	defer l.mu.Unlock()
-	l.logger.Out = l.asyncWriter
+	l.logger.Out = l.logWriter
 	return nil
 }