@ryarnyah 2018-04-08
Replaced with https://github.com/containous/traefik/pull/3152